Basic Pause Implementation
Countdown Timer now is clamped to 0 meaning that it will never show a
negative time.

Pause function created when minutes are equal to 0 and seconds are less
than 1 second.  Currently the game has a resume function (just in case)
though it isn't implemented yet.

Next Step: Create a Restart Level UI that allows the player to reset the
level if time runs out.  This UI may also be able to show certain stats
at a future time.